Changed the name of the HAL to prevent conflict with ESP IDF HAL.
